backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2

# Database connections
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0

# Vector database and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# AI/LLM
transformers==4.35.2
torch==2.1.1
openai==1.3.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
